+++ test/manual/hicache/test_disaggregation_hicache.py
@@ -8,7 +8,7 @@ from typing import Dict
 import requests
 
 from sglang.bench_serving import get_tokenizer
-from sglang.test.server_fixtures.test_disaggregation_utils import TestDisaggregationBase
+from sglang.test.server_fixtures.disaggreation_fixture import TestDisaggregationBase
 from sglang.test.test_utils import (
     DEFAULT_MODEL_NAME_FOR_TEST,
     DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
