+++ test/manual/piecewise_cudagraph/test_disaggregation_piecewise_cuda_graph.py
@@ -2,7 +2,7 @@ import unittest
 from types import SimpleNamespace
 
 from sglang.test.few_shot_gsm8k import run_eval as run_eval_few_shot_gsm8k
-from sglang.test.server_fixtures.test_disaggregation_utils import TestDisaggregationBase
+from sglang.test.server_fixtures.disaggreation_fixture import TestDisaggregationBase
 from sglang.test.test_utils import (
     DEFAULT_MODEL_NAME_FOR_TEST,
     DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
