+++ test/srt/test_disaggregation_hybrid_attention.py
@@ -2,7 +2,7 @@ import unittest
 from types import SimpleNamespace
 
 from sglang.test.few_shot_gsm8k import run_eval as run_eval_few_shot_gsm8k
-from sglang.test.server_fixtures.test_disaggregation_utils import TestDisaggregationBase
+from sglang.test.server_fixtures.disaggreation_fixture import TestDisaggregationBase
 from sglang.test.test_utils import (
     DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
     popen_launch_pd_server,
