+++ test/srt/test_disaggregation_pp.py
@@ -3,7 +3,7 @@ import unittest
 from types import SimpleNamespace
 
 from sglang.test.few_shot_gsm8k import run_eval
-from sglang.test.server_fixtures.test_disaggregation_utils import TestDisaggregationBase
+from sglang.test.server_fixtures.disaggreation_fixture import TestDisaggregationBase
 from sglang.test.test_utils import (
     DEFAULT_MODEL_NAME_FOR_TEST,
     DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
